Really fix warnings in stringfeatures.

+++ src/libshogun/features/StringFeatures.h
@@ -1228,7 +1228,7 @@ template <class ST> class CStringFeatures : public CFeatures
 		 */
 		virtual TString<ST>* get_features(int32_t& num_str, int32_t& max_str_len)
 		{
-			if(m_subset_idx)
+			if (m_subset_idx)
 				SG_ERROR("not possible on subset");
 
 			num_str=num_vectors_total;
@@ -1302,36 +1302,36 @@ template <class ST> class CStringFeatures : public CFeatures
 
 			// header shogun v0
 			char id[4];
-			if(fread(&id[0], sizeof(char), 1, file)!=sizeof(char))
+			if (fread(&id[0], sizeof(char), 1, file)!=sizeof(char))
 				SG_ERROR("failed to read header");
 			ASSERT(id[0]=='S');
-			if(fread(&id[1], sizeof(char), 1, file)!=sizeof(char))
+			if (fread(&id[1], sizeof(char), 1, file)!=sizeof(char))
 				SG_ERROR("failed to read header");
 			ASSERT(id[1]=='G');
-			if(fread(&id[2], sizeof(char), 1, file)!=sizeof(char))
+			if (fread(&id[2], sizeof(char), 1, file)!=sizeof(char))
 				SG_ERROR("failed to read header");
 			ASSERT(id[2]=='V');
-			if(fread(&id[3], sizeof(char), 1, file)!=sizeof(char))
+			if (fread(&id[3], sizeof(char), 1, file)!=sizeof(char))
 				SG_ERROR("failed to read header");
 			ASSERT(id[3]=='0');
 
 			//compression type
 			uint8_t c;
-			if(fread(&c, sizeof(uint8_t), 1, file)!=sizeof(uint8_t))
+			if (fread(&c, sizeof(uint8_t), 1, file)!=sizeof(uint8_t))
 				SG_ERROR("failed to read compression type");
 			CCompressor* compressor= new CCompressor((E_COMPRESSION_TYPE) c);
 			//alphabet
 			uint8_t a;
 			delete alphabet;
-			if(fread(&a, sizeof(uint8_t), 1, file)!=sizeof(uint8_t))
+			if (fread(&a, sizeof(uint8_t), 1, file)!=sizeof(uint8_t))
 				SG_ERROR("failed to read compression alphabet");
 			alphabet=new CAlphabet((EAlphabet) a);
 			// number of vectors
-			if(fread(&num_vectors, sizeof(int32_t), 1, file)!=sizeof(int32_t))
+			if (fread(&num_vectors, sizeof(int32_t), 1, file)!=sizeof(int32_t))
 				SG_ERROR("failed to read compression number of vectors");
 			ASSERT(num_vectors>0);
 			// maximum string length
-			if(fread(&max_string_length, sizeof(int32_t), 1, file)!=sizeof(int32_t))
+			if (fread(&max_string_length, sizeof(int32_t), 1, file)!=sizeof(int32_t))
 				SG_ERROR("failed to read maximum string length");
 			ASSERT(max_string_length>0);
 
@@ -1342,11 +1342,11 @@ template <class ST> class CStringFeatures : public CFeatures
 			{
 				// vector len compressed
 				int32_t len_compressed;
-				if(fread(&len_compressed, sizeof(int32_t), 1, file)!=sizeof(int32_t))
+				if (fread(&len_compressed, sizeof(int32_t), 1, file)!=sizeof(int32_t))
 					SG_ERROR("failed to read vector length compressed");
 				// vector len uncompressed
 				int32_t len_uncompressed;
-				if(fread(&len_uncompressed, sizeof(int32_t), 1, file)!=sizeof(int32_t))
+				if (fread(&len_uncompressed, sizeof(int32_t), 1, file)!=sizeof(int32_t))
 					SG_ERROR("failed to read vector length uncompressed");
 
 				// vector raw data
@@ -1355,7 +1355,7 @@ template <class ST> class CStringFeatures : public CFeatures
 					features[i].string=new ST[len_uncompressed];
 					features[i].length=len_uncompressed;
 					uint8_t* compressed=new uint8_t[len_compressed];
-					if(fread(compressed, len_compressed, 1, file)!=len_compressed)
+					if (fread(compressed, len_compressed, 1, file)!=(size_t) len_compressed)
 						SG_ERROR("failed to read compressed data");
 					uint64_t uncompressed_size=len_uncompressed;
 					uncompressed_size*=sizeof(ST);
@@ -1374,7 +1374,7 @@ template <class ST> class CStringFeatures : public CFeatures
 					feat32ptr[0]=(int32_t) len_compressed;
 					feat32ptr[1]=(int32_t) len_uncompressed;
 					uint8_t* compressed=(uint8_t*) (&features[i].string[offs]);
-					if(fread(compressed, len_compressed, 1, file)!=len_compressed)
+					if (fread(compressed, len_compressed, 1, file)!=(size_t) len_compressed)
 						SG_ERROR("failed to read uncompressed data");
 				}
 			}
